RavenDB-5854 - fix failing tests - changed the total installed memory to double - added the option to provide the license inside a file - allow to skip logging lease license errors

+++ src/Raven.Client/ServerWide/Commands/GetNodeInfoCommand.cs
@@ -12,7 +12,7 @@ namespace Raven.Client.ServerWide.Commands
         public string Certificate;
         public string ClusterStatus;
         public int NumberOfCores;
-        public int InstalledMemoryInGb;
+        public double InstalledMemoryInGb;
         public double UsableMemoryInGb;
     }
